add function to read file
! 此次更新的程式碼仍屬於草稿階段，後續還需要再整理。
! 設定檔案使用 `# ` 作為表示「標題」的符號，後續直到下一個標題間的內容，則為此標題底下的值。

* 結構改寫：將具體功能分離至 `Generator.js` ，僅在 `niceidea.js` 保留使用者介面邏輯。
* 將功能分割為 `Generator` 與 `Random` 兩個類，分別負責功能邏輯與產生隨機內容。
* 將固定值如音符列表、調式列表等內容，分割至 `Assert.js` 的 `Assert` 中。

+ 添加了部份類型註解。
+ 添加了讀取設定檔的功能與相關介面。
+ 添加了一個設定檔案範本。
